requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Frontend
gradio>=5.0.0

# AI Providers
google-generativeai>=0.5.0
openai>=1.30.0
anthropic>=0.25.0

# Job Queue
celery>=5.3.0
redis>=5.0.0

# Vector Store
faiss-cpu>=1.8.0
langchain>=0.2.0
langchain-community>=0.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
aiohttp>=3.9.0
PyYAML>=6.0
numpy>=1.26.0

# Audio
ffmpeg-python>=0.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
